My math Program
This is a rather fun program that helped me realize the power of functions and how to integrate functions to make a working menu.
Not only was it a fun project but this also helps me with my math course. Every day I do homework I open this up and have it out.
I hope that you can use this program to help you as much as it helps me.